memtest86+: Add coreboot branding and version number
So that this doesn't appear to be a default memtest86+, add the text
'coreboot' followed by a version to the memtest86+ version string.

The intention for this is mainly so that any issues that show up
in this come to coreboot for debug, and don't bother the original
devs with issues that crop up due to our modifications.

- Update the code so that the 'M' character in memtest doesn't
flash anymore.
- Put the + character in the version string.
- Move the version string into config.h so it's easy to locate
and update.

Change-Id: I1b374ad4d9e3de91a703be88aee5a3f5bf16fd25
Signed-off-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://review.coreboot.org/13830
Reviewed-by: Ben Gardner <gardner.ben@gmail.com>