memtest86+: Update linuxbios names to coreboot
This does part of a patch that Stefan wrote for memtest86-3.5,
to update the linuxbios names to coreboot.

https://www.flashrom.org/pipermail/coreboot/2010-January/055363.html

coreboot_tables.h was updated from coreboot's copy at
commonlib/coreboot_tables.h.  Stefan's original patch changed the
names of the functions and data structures to cb_xxxx, but since
that's not done in the actual coreboot codebase, I left them named
lb_xxxx here.

Change-Id: Ib14bf84ef0527d8277b52461a7b98c0f3566cfb9
Signed-off-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://review.coreboot.org/13944
Tested-by: build bot (Jenkins)
Reviewed-by: Ben Gardner <gardner.ben@gmail.com>