Memtest86+: fix missing break in while statement
Coverity spotted a case of missing breaks in a switch.
The net effect is that instead of showing a CPU type guess of "586" or
"686", it shows "Unidentified Processor".

I'm not aware of any CPUs that would hit this code. This switch statement
may be effectively dead code, but we may as well have correct dead code.

Change-Id: I93349cf1ff9b9366610428c8e8ae3ae60d23b4a4
Signed-off-by: Ben Gardner <gardner.ben@gmail.com>
Found-by: Coverity (75579, 75580)
Reviewed-on: https://review.coreboot.org/14196
Tested-by: build bot (Jenkins)
Reviewed-by: Martin Roth <martinroth@google.com>