Update makefile to allow using coreboot's toolchain
coreboot sets the AS and CC variables when building memtest86+, but they
currently get reset by the Memtest86+ makefile.  This change keeps them
set to the values used by coreboot.

Add the flag '-32' after AS is set so that it still gets added if the
AS variable is set somewhere else.

Change-Id: Id211a21a87a0fee7e17170a2d75bad3237d910e0
Signed-off-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://review.coreboot.org/14458
Tested-by: build bot (Jenkins)
Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>